--- BE_SistemaViviendas/components/Atuh/login.model.js
@@ -1,7 +1,0 @@
-
-var loginModel = {
-    userName: String,
-    password: String
-}
-
-module.exports = loginModel
