+++ BE_SistemaViviendas/components/Atuh/login.query.js
@@ -1,0 +1,9 @@
+//QUERYS    
+
+function queryLogin(userLogin) {
+    return query = `SELECT * FROM users WHERE userName = '${userLogin.userName}' and password = '${userLogin.password}'`;
+}
+
+module.exports = {
+    queryLogin
+}
