+++ BE_SistemaViviendas/config/config.js
@@ -1,13 +1,1 @@
-// coneccion a base de datos
-var mysql = require('mysql');
-
-var con = mysql.createConnection({
-    host: "localhost",
-    port: 3301,
-    user: "root",
-    password: "Nicole2630:",
-    insecureAuth : false,
-    database: 'db_roofit'
-  });
- 
-module.exports = con
+module.exports.SEED = '@este-es-un@seed-roofIT';
